pyproject.toml
[project]
name = "linkedin-poster"
version = "0.1.0"
description = "LinkedIn posting automation"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "python-dotenv",
    "google-generativeai",
    "tavily-python",
    "python-telegram-bot",
    "schedule",
    "openai"
]
